Move fields in new files
